+++ convert_type.go
@@ -1,10 +1,10 @@
-package fsrpl
+package main
 
 import "time"
 
 const fsTimeLayout = "2006-01-02T15:04:05Z"
 
-// InterpretationEachValueForTime :
+// InterpretationEachValueForTime convert string to time.Time
 func InterpretationEachValueForTime(mps map[string]interface{}) map[string]interface{} {
 	for k, v := range mps {
 		vs, ok := v.(string)
