+++ version.go
@@ -1,0 +1,5 @@
+package main
+
+var (
+	version = "0.1.0"
+)
